// src/app/aliens/models/soldier.ts
export interface Soldier {
  id: number;
  name: string;
  rank: string;
  nationality: string;
  timeUnits: number;
  energy: number;
  health: number;
  bravery: number;
  reactions: number;
  firingAccuracy: number;
  throwingAccuracy: number;
  strength: number;
  psiSkill: number;
  psiStrength: number;
  meleeAccuracy: number;
  armourType: string;
  missions: number;
  killed: number;
  picture: number;
}